style: OwnerBoard 패키지와 OwnerBoardComment 패키지 병합
