+++ src/main/java/com/study/petory/domain/ownerBoard/dto/request/OwnerBoardCommentUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.study.petory.domain.ownerBoardComment.dto.request;
+package com.study.petory.domain.ownerBoard.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
