+++ src/main/java/com/study/petory/domain/ownerBoard/repository/OwnerBoardCommentRepository.java
@@ -1,4 +1,4 @@
-package com.study.petory.domain.ownerBoardComment.repository;
+package com.study.petory.domain.ownerBoard.repository;
 
 import java.util.List;
 
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.study.petory.domain.ownerBoardComment.entity.OwnerBoardComment;
+import com.study.petory.domain.ownerBoard.entity.OwnerBoardComment;
 
 public interface OwnerBoardCommentRepository extends JpaRepository<OwnerBoardComment, Long> {
