+++ src/main/java/com/study/petory/domain/ownerBoard/service/OwnerBoardServiceImpl.java
@@ -16,9 +16,9 @@ import com.study.petory.domain.ownerBoard.dto.response.OwnerBoardGetResponseDto;
 import com.study.petory.domain.ownerBoard.dto.response.OwnerBoardUpdateResponseDto;
 import com.study.petory.domain.ownerBoard.entity.OwnerBoard;
 import com.study.petory.domain.ownerBoard.repository.OwnerBoardRepository;
-import com.study.petory.domain.ownerBoardComment.dto.response.OwnerBoardCommentGetResponseDto;
-import com.study.petory.domain.ownerBoardComment.entity.OwnerBoardComment;
-import com.study.petory.domain.ownerBoardComment.repository.OwnerBoardCommentRepository;
+import com.study.petory.domain.ownerBoard.dto.response.OwnerBoardCommentGetResponseDto;
+import com.study.petory.domain.ownerBoard.entity.OwnerBoardComment;
+import com.study.petory.domain.ownerBoard.repository.OwnerBoardCommentRepository;
 import com.study.petory.domain.user.entity.User;
 import com.study.petory.domain.user.repository.UserRepository;
 import com.study.petory.exception.CustomException;
